chore: move remaining files and add missing imports
