+++ apps/frontend-manage/package.json
@@ -70,6 +70,7 @@
     "@types/node": "^20.16.1",
     "@types/react": "^18.3.11",
     "@types/react-dom": "^18.3.1",
+    "@types/react-tagcloud": "^2.3.2",
     "autoprefixer": "~10.4.20",
     "cross-env": "~7.0.3",
     "cssnano": "~6.0.1",
@@ -77,6 +78,7 @@
     "eslint-config-next": "~15.0.0",
     "postcss": "~8.4.47",
     "postcss-import": "~16.1.0",
+    "react-tagcloud": "~2.3.3",
     "swc-loader": "~0.2.6",
     "tailwindcss": "~3.4.14",
     "tailwindcss-radix": "~3.0.5",
